requirements.txt
streamlit
python-docx
PyMuPDF
sentence-transformers
scikit-learn
groq
python-dotenv
